requirements.txt
python-dotenv>=1.0.0
PyYAML>=6.0.1
ruamel.yaml>=0.18.0
python-frontmatter>=1.1.0
chromadb>=0.5.0
openai>=1.30.0
google-genai>=1.0.0
watchdog>=4.0.0
requests>=2.31.0
rich>=13.7.0
click>=8.1.7
